+++ pgxpool/rows.go
@@ -1,7 +1,7 @@
 package pgxpool
 
 import (
-	"github.com/jackc/pgconn"
+	"github.com/ethanpailes/pgconn"
 	"github.com/jackc/pgproto3/v2"
 	"github.com/jackc/pgx/v4"
 )