+++ pgxpool/tx.go
@@ -3,7 +3,7 @@ package pgxpool
 import (
 	"context"
 
-	"github.com/jackc/pgconn"
+	"github.com/ethanpailes/pgconn"
 	"github.com/jackc/pgx/v4"
 )