+++ rows.go
@@ -7,7 +7,7 @@ import (
 
 	errors "golang.org/x/xerrors"
 
-	"github.com/jackc/pgconn"
+	"github.com/ethanpailes/pgconn"
 	"github.com/jackc/pgproto3/v2"
 	"github.com/jackc/pgtype"
 )